+++ connecteur/as@lae-rest/AsalaeREST.class.php
@@ -188,21 +188,24 @@ class AsalaeREST extends SAEConnecteur {
 	 * @throws Exception
 	 */
 	public function getReply($id_transfert) {
-		$org = $this->originatingAgency;
-		$result = $this->getWS(
-			"/sedaMessages/sequence:ArchiveTransfer/message:ArchiveTransferReply/originOrganizationIdentification:$org/originMessageIdentifier:"
-            .urlencode($id_transfert),
+		if (! $id_transfert){
+			throw new UnrecoverableException("L'identifiant du transfert n'a pas été trouvé");
+		}
+
+		return $this->getWS(
+			sprintf(
+				"/sedaMessages/sequence:ArchiveTransfer/message:ArchiveTransferReply/originOrganizationIdentification:%s/originMessageIdentifier:%s",
+				$this->originatingAgency,
+				urlencode($id_transfert)
+			),
 			"application/xml"
 		);
-		//WTF : ca ne peut jamais arriver ce truc !
-		if (!$result){
-			$this->last_error_code = 8;
-			return false;
-		}
-		return $result;	
 	}
 	
 	public function getURL($cote) {
+		if (empty($this->url)){
+			return $cote;
+		}
 		$tab = parse_url($this->url);
 		return "{$tab['scheme']}://{$tab['host']}/archives/viewByArchiveIdentifier/$cote";
 	}