+++ test/PHPUnit/connecteur-type/sae/SAEVerifierTest.php
@@ -2,13 +2,17 @@
 
 class SAEVerifierTest extends PastellTestCase {
 
+	use CurlUtilitiesTestTrait;
 
 	/**
 	 * @throws NotFoundException
 	 */
 	public function testVerifier(){
 
-		$this->mockCurl();
+		$this->mockCurl([
+			'/sedaMessages/sequence:ArchiveTransfer/message:Acknowledgement/originOrganizationIdentification:/originMessageIdentifier:15ef78ef665a8777c33d1125783707f8dfb190f82869dc9248e46c5ed396d70b_1542893421'
+			=> file_get_contents(__DIR__."/fixtures/acuse-de-reception-asalae.xml")
+		]);
 
 		$id_ce = $this->createConnector('as@lae-rest',"Asalae")['id_ce'];
 		$this->associateFluxWithConnector($id_ce,"actes-generique","SAE");
@@ -35,33 +39,6 @@ class SAEVerifierTest extends PastellTestCase {
 		);
 	}
 
-	private function mockCurl(){
-		$curlWrapper = $this->getMockBuilder('CurlWrapper')
-			->disableOriginalConstructor()
-			->getMock();
 
-		$curlWrapper->expects($this->any())
-			->method('get')
-			->will($this->returnCallback(function($url){
-				if ($url == '/sedaMessages/sequence:ArchiveTransfer/message:Acknowledgement/originOrganizationIdentification:/originMessageIdentifier:15ef78ef665a8777c33d1125783707f8dfb190f82869dc9248e46c5ed396d70b_1542893421'){
-					return file_get_contents(__DIR__."/fixtures/acuse-de-reception-asalae.xml");
-				}
-				throw new UnrecoverableException("Appel à une URL inatendue $url");
-			}));
-
-		$curlWrapper->expects($this->any())
-			->method('getHTTPCode')
-			->willReturn(200);
-
-		$curlWrapperFactory = $this->getMockBuilder('CurlWrapperFactory')
-			->disableOriginalConstructor()
-			->getMock();
-
-		$curlWrapperFactory->expects($this->any())
-			->method('getInstance')
-			->willReturn($curlWrapper);
-
-		$this->getObjectInstancier()->setInstance(CurlWrapperFactory::class,$curlWrapperFactory);
-	}
 
 }